NPT-380 Support for copy over of const definitions from DSL to generated file
Currently const() definitions are not copied over by nexus compiler to generated files.
This means, applications do not have access to the enums/consts when trying to use the constants.
compiler should copy over constants, as it is, to the generated types.

Now, with this changes compiler will copy over constants from DSL to the generated files.