NPT-369 Make api-gw singleton nodes aware
Singleton nodes are nodes which can only have 'default' as a name.
Because of that there's no need to add name of this node as a query
param, default can be assumed. If name is different than default
validator will throw an error. If name is not passed, api-gw will
use 'default' for singleton parents and for singleton node.