NPT-400 Shim layer support for custom user status
Need to add API in shim layer to Set, Get and Clear the user defined status.

This changes is to add API in shim layer to Set, Get and Clear the user defined status.
Set, Get and Clear operations on status can be done if object instance or object chainer is given.