TSM-7650: Add annotation "x-nexus-short-name" to support short names (tag bumped)
Add support for shortnames to tsm-cli get

To support enhancement requests from PM, we need to add a short name to the tsm-cli get operations. For this, we need to have one more annotation as "x-nexus-short-name" for the supported Get API calls.

The "x-nexus-short-name" included in the GET APIs

See merge request nsx-allspark_users/nexus-sdk/cli!126