NPT-482 Define nexus types for graphql custom query and filter
Users will need to use nexus DSL types to specify custom queries
and filters that are to be associated with a nexus node. So there
is a need to standardize the DSL format that the users can use.

This commit creates the basic types and spec that standardizes the
spec with which the user can specify custom graphql queries and filters.

Closes NPT-482

See merge request nsx-allspark_users/nexus-sdk/nexus!41