NPT-579 Add hierarchy and custom args to GetMetrics API
We want to extend GetMetrics API with additional params.

Hierarachy will be sent as a part of query, Custom Args will be added
to query params.
In GetMetrics arguments can be a subset of well-known MetricsArg
arguments (`Metric`, `StartTime`, `EndTime`, `TimeInterval`) or
additional custom arguments which will be translated into a `UserProvidedArgs` map.