NPT-599 Increase resolver count for custom queries
In graphql schema template we depend on resolver count being greater than 0,
if it's zero schema for given node is not generated. For now resolver
count was increased only if there was a child/link node defined, so when
there weren't any child or node count was 0, so there schema was not
resolved properly for custom queries.

Now increase resolver count also for custom queries so template is
resolved properly when there is no child or link, but only custom
queries in given node.