NPT-647 Remove host,port env check is they are empty
Connect-controller ensures that the host env is not empty. If not empty, then process the replication-configs that are of interest to the desired destination, by comparing env with endpoint object linked to the config.

If connector is standalone, then host could be empty. In this case, process all the replication-configs.

This commit addresses both the cases.

Closes NPT-647

See merge request nsx-allspark_users/nexus-sdk/connector!39