[NPT-532] In Ubuntu machine, the contents of datamodel build directory are created as root, resulting in subsequence build failures
When running datamodel build, the contents of the build directory are being created as a root user (as the files are generated inside the docker container). This results in inablity to run build again, as the makefile outside the container tries to cleanup build directory.

Run chown when compiler finishes generating files

See merge request nsx-allspark_users/nexus-sdk/datamodel-templates!37