NPT-644: Compare nexus node spec (#66)
NPT-644: Compare nexus node spec

When user changes nexus node source files, sometimes the change can be backward incompatible. Nonetheless, user should be able to upgrade model and see the incompatible changes.

This change adds library in common-library that allows to see if there are any changes in 2 bytes arrays and return answer as boolean and human-readable report (in buffer type, that can be printed).