NPT-711 Change nexus datamodel api group to nexus.vmware.com
The nexus datamodel api group is currently configured to be nexus.org.
The ask is to change this group name to nexus.vmware.com.
This group naming is inline with other group names in the Tanzu product line.

This commit changes the group name to nexus.vmware.com

Closes NPT-711

See merge request nsx-allspark_users/nexus-sdk/nexus-runtime-manifests!39